fsl_qbman: fix mismatched asm register operator widths in safe_copy_dqrr()
We are requesting an "and" and an "add" operation between registers of
mismatched sizes: "*s64" is 64-bit, while "res" and "zero" are 32-bit.
Promote "res" and "zero" to 64 bits, to get rid of this clang-16 build
failure:

drivers/staging/fsl_qbman/qman_high.c:1160:26: error:
value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
                             : [result] "=r" (res)
                                              ^
drivers/staging/fsl_qbman/qman_high.c:1159:21: note: use constraint modifier "w"
                asm volatile("and %[result], %[in1], %[in2]"
                                  ^~~~~~~~~
                                  %w[result]
drivers/staging/fsl_qbman/qman_high.c:1161:22: error:
value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
                             : [in1] "r" (zero), [in2] "r" (*s64)
                                          ^
drivers/staging/fsl_qbman/qman_high.c:1159:32: note: use constraint modifier "w"
                asm volatile("and %[result], %[in1], %[in2]"
                                             ^~~~~~
                                             %w[in1]
drivers/staging/fsl_qbman/qman_high.c:1170:22: error:
value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
                             : [in1] "r" (res), [in2] "r" (s64)
                                          ^
drivers/staging/fsl_qbman/qman_high.c:1168:32: note: use constraint modifier "w"
                asm volatile("add %[result], %[in1], %[in2]"
                                             ^~~~~~
                                             %w[in1]

+++ drivers/staging/fsl_qbman/qman_high.c
@@ -1151,7 +1151,7 @@ static inline void safe_copy_dqrr(struct qm_dqrr_entry *dst,
 	*d64 = *s64;
 #if defined(CONFIG_ARM) || defined(CONFIG_ARM64)
 	{
-		u32 res, zero = 0;
+		u64 res, zero = 0;
 		/* Create a dependancy after copying first bytes ensures no wrap
 		   transaction generated to QBMan */
 		/* Logical AND the value pointed to by s64 with 0x0 and
